requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cloudinary==1.29.0
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-heroku==0.3.1
django-location-field==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
channels==3.0.4
channels-redis==3.3.1